add shape checker and split attribute
